Drop "float32" and "float64" types
They would be even more useless than "float" and "double" in practice.
Also rename "portable_binfloat" to just "binfloat", because this shall
be the floating-point format of choice, and thus not be much longer to
type than "float" or "double".

Also add decimal_N types for simple fixed-point decimal values, such as
monetary amounts with decimal_2.