Rename portable floating type to "pfloat"
It shall be short! Also change the definitions for the decimal types.

"decimal" is now similar to a "pfloat" but always uses radix 10. Also
extend "pfloat" to also support radix 10, although radix 2 is the
default in a "pfloat" and radix 10 will produce longer encodings.

Introduce a new type "dfix_4" which can selectively store 0, 1, 2 or 4
fractional decimal digits exactly and space-efficient.

Finally, add "dfix_1" and "dfix_2" as decimal fixed point types with
exactly 1 and 2 fractional digits, respectively.